Add edit/play buttons
