+++ index.html
@@ -7,7 +7,10 @@
 <body>
 
 <div id='editor-container'>
-
+	<canvas id='editor'></canvas>
+	<canvas id='game' style='display: none'></canvas>
+	<div class='button button-edit' id='button-edit'>edit</div>
+	<div class='button button-play' id='button-play'>play</div>
 </div>
 
 <div id='sidebar'>
